requirements.txt
# Core Dependencies
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0

# CARLA Simulator
carla==0.9.15

# Vision & Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
h5py>=3.9.0

# Configuration Management
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# Logging & Monitoring
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.66.0

# Utilities
einops>=0.7.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development Tools
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
